fix: keep the browser freeze wrapper out of napi's generated browser.js (Codex P2)
The Codex finding's literal claim — that `napi build` overwrites the `main`
(`wasm-tools.js`) and `types` (`wasm-tools.d.ts`) wrappers because those fields
are its default output paths — is FALSE for @napi-rs/cli 3.6.1. Verified two
ways:
  * source: `napi build` writes the JS binding to `options.jsBinding ?? 'index.js'`
    and the d.ts to `this.options.dts ?? 'index.d.ts'` — both HARDCODED, never
    read from package.json `main`/`types`.
  * empirical: `pnpm build` twice — `wasm-tools.js`/`wasm-tools.d.ts` md5 unchanged,
    `git status` shows neither. The wrappers survive.

But the finding pointed at a REAL adjacent bug in a DIFFERENT file. `napi build`
DOES own `browser.js`: it regenerates it to a bare `export * from
'…-wasm32-wasi'` on every build (filename hardcoded at cli `index.js:1628`), so
the hand-authored deep-freeze loop that lived in `browser.js` was silently erased
— proven: after `pnpm build`, `browser.js`'s freeze loop was gone. The
shared-singleton protection on the wasm/browser target was an illusion that any
build (incl. CI/publish) would destroy.

Fix mirrors the native pattern (main -> wasm-tools.js wraps generated index.js):
move the browser freeze wrapper to a hand-authored `wasm-tools.browser.js` and
point `"browser"` at it. napi keeps regenerating the throwaway `browser.js`;
the wrapper is never touched. Proven durable: two rebuilds leave
`wasm-tools.browser.js` byte-identical while only `browser.js` is rewritten
(to the same bare form). 437 tests pass on release; the native freeze test is green.
